add:arquivos de funções de clientes e extra
