+++ moduloExtra/funcoesExtra.h
@@ -1,0 +1,2 @@
+void telaAddValor(void);
+void telaConfirmacao(void);
