+++ moduloProduto/funcoesProduto.h
@@ -3,3 +3,4 @@ void telaCadastroProduto(void);
 void telaPesquisarProduto(void);
 void telaApagarProduto(void);
 char telaAtualizarProduto(void);
+void navegacaoMenuProduto(void);
